# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SpeakSum API
  version: 2.0.0
  description: |
    SpeakSum 当前运行中的接口契约。

    当前主模型为：
    - content
    - summary_text
    - quotes
    - domains
    - domain graph
servers:
  - url: http://127.0.0.1:8000
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /api/v1/upload:
    post:
      summary: Upload a content file
      tags: [Upload]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                source_type:
                  type: string
                  enum: [meeting_minutes, other_text]
                  default: meeting_minutes
                provider:
                  type: string
                  description: Model provider to use for processing
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadAccepted'
  /api/v1/upload/{task_id}/status:
    get:
      summary: Get task status
      tags: [Upload]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingStatus'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UploadAccepted:
      type: object
      properties:
        task_id:
          type: string
        content_id:
          type: string
        status:
          type: string
          enum: [pending]
    ProcessingStatus:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
        stage:
          type: string
          nullable: true
        percent:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        content_id:
          type: string
          nullable: true
        meeting_id:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
...